Add EthbridgeService and tidy up
