+++ ui/core/src/api/EthbridgeService/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./EthbridgeService";
+export * from "./EthbridgeService";
