+++ ui/core/src/api/utils/SifClient/SifClient.ts
@@ -1,0 +1,3 @@
+import { SigningCosmosClient } from "@cosmjs/launchpad";
+
+export class SifClient extends SigningCosmosClient {}
